requirements.txt
agate==1.9.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.1
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.12.1
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
certifi==2026.1.4
charset-normalizer==3.4.4
chess-guru==0.1.2
chesscom-guru==0.1.5
click==8.3.1
colorama==0.4.6
coloredlogs==14.0
daff==1.4.2
dagster_shared==1.12.12
dagster-dbt==0.28.12
dagster-graphql==1.12.12
dagster-pipes==1.12.12
dagster-webserver==1.12.12
dagster==1.12.12
dbt-adapters==1.16.3
dbt-common==1.27.1
dbt-core-interface==1.1.7
dbt-core==1.9.1
dbt-extractor==0.6.0
dbt-osmosis==1.2.2
dbt-postgres==1.9.1
dbt-protos==1.0.422
dbt-semantic-interfaces==0.7.5
deepdiff==7.0.1
distro==1.9.0
docstring_parser==0.17.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
gql==4.0.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.1
grpcio-health-checking==1.71.2
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
isodate==0.6.1
Jinja2==3.1.6
jiter==0.12.0
jsonschema-specifications==2025.9.1
jsonschema==4.26.0
leather==0.4.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mashumaro==3.14
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.7.0
mysql-mimic==2.6.4
networkx==3.6.1
openai==2.15.0
ordered-set==4.1.0
orjson==3.11.5
packaging==26.0
parsedatetime==2.6
pathlib_abc==0.5.2
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
protobuf==5.29.5
psutil==7.2.1
psycopg2-binary==2.9.11
pydantic_core==2.41.5
pydantic==2.12.5
Pygments==2.19.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pywin32==311
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
requests-toolbelt==1.0.0
requests==2.32.5
rich==14.3.1
rpds-py==0.30.0
ruamel.yaml.clib==0.2.15
ruamel.yaml==0.18.17
setuptools==80.10.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowplow-tracker==1.1.0
SQLAlchemy==2.0.46
sqlglot==28.0.0
sqlglotrs==0.7.3
sqlparse==0.5.5
starlette==0.52.1
structlog==25.5.0
tabulate==0.9.0
text-unidecode==1.3
tomli==2.4.0
tomlkit==0.14.0
toposort==1.10
tqdm==4.67.1
typer==0.21.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
universal_pathlib==0.3.8
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
yarl==1.22.0
zipp==3.23.0
